modules: always #include "mw-modules.h"
The ensures the public functions declarations are consistent with the
API, and prevents accidental variable/function name collisions. Some
variables or parameters had to be renamed, including some variables in
the preprocessor header. The renaming tries to keep an understandable
name, but may sometimes look strange. A global rewriting would sometimes
help.