modules/image: remove unused parameters and fix signedness
Unused parameters in static functions are removed. Unused parameters in
public functions or X11 callbacks can't be removed because they impact
the command-line or are required even if unused, so a "foo=foo"
workaround is inserted to fool the compiler.

Stick to the structure definitions about structure members signedness
(some of them should be changes, images width/heith is unsigned in fact,
but changing it would break the API).