modules: rename shadowed/shadowing objects
Some local variables had the same name as some globally accessible
objects (variables or functions). This is bad, because confusing, and it
makes the globally accessible objects unusable in the local context. So
the local variables have been renamed. The renaming tries to keep an
understandable name, but may sometimes look strange. A global rewriting
would sometimes help.