libmw-x11: don't #include <unistd.h>, use POSIX nanosleep() for the moment
We want to reduce our dependencies to ANSI C only, so the OS specific
headers, such as unistd.h, can't be used anymore.

Unfortunately, no portable equivalent of usleep() seems to be available,
eithen in the ANSI C or X11 libraries. For the moment, we use the POSIX
nanosleep() implementation from time.h, thus limiting the libmw-x11
compilation to POSIX systems. Exploring portable alternatives is
desirable, vlc msleep() for example.
http://www.videolan.org/developers/vlc/doc/doxygen/html/mtime_8c.html