mwp: don't #include <sys/stat.h>
We want to reduce our dependencies to ANSI C only, so the OS specific
headers, such as sys/stat.h, can't be used anymore. So, the copyright
information in the documentation output will be 1993-XXXX where XXXX is
the year of compilation. This will match the latest release for each
release. Anyway, relying on the file's mtime was not safe because of
possible file copy. The exact last update time of the files can be
provided by the source control system, or as a comment in the header.