libmw/mwio.c: only #include ANSI C headers
We want to reduce our dependencies to ANSI C only, so the OS specific
headers, such as unistd.h, fcntl.h or sys/stat.h, can't be used anymore.

As there is no concept of "directory" in ANSI C, we drop the recursive
exploration of the filesystem. Instead, a file foo can be:
1- in the current working directory
2- in $MY_MEGAWAVE2/data/<mwgroup>/
3- in $MY_MEGAWAVE2/data/
4- in $MEGAWAVE2/data/<mwgroup>/
5- in $MEGAWAVE2/data/
6- in any part of $MEGAWAVE2_DATAPATH ($MEGAWAVE2_DATAPATH is a
   colon-separated list of directories, like the usual $PATH)

Maybe /usr/share/megawave3, /usr/local/share/megawave3 and
~/.config/megawave3 could be added later.