libmw: avoid printf() with non-literal format
Data/format mismatch is a source of run-time crash and hard to debug.
GCC can check it at compile-time, if the format string is literal, known
at compile time (ie not a variable).

When possible, format variables have been replaced by explicit strings.
If it was not possible for the moment, a special #pragma is added to
exclude these functions from GCC strict checks.

The movie frame name pattern is changed; now it is always written with
five digits. checkmodules.sh had to be updated in consequence.