libmw-commandline: no more call_proto() and call_fsum()
call_proto() and call_fsum() were seldom used functions, and their
interest (printing the function prototype) is uncertain. In relation
with the removal of protobuf in mwp, this cleanup helps keep the code
simple and safe.

Mwiline definiton is now included in libmw-commandline, not libmw. To
ensure libmw and libmw-cmdline separation, mwerror(USAGE,...) doesn't
print the module usage anymore, but only suggests to use the --help
option.