modules: use the Mersenne Twister PRNG
The libc standard pseudo-random number generation usualy have poor
randomness. Instead of it, we prefer a better implementation, based on
the Mersenne Twister PRNG. As far as we know, this PRNG is supposed to
be of very high quality, fast, with a very long period (2^19937 - 1),
high order dimensional equidistribution, and very good statistical
randomness.