modules/curve: fix unitialized variables and unhandled return values
Compiler optimisation mode enables deep code flow analysis, and catches
potential unitialized variables. They have been fixed by a naive
initialization. Serious code study would be better and is likely to
catch serious bugs.

Return values should also be checked; safe and sound handling of scanf()
is not trivial, so we use a dummy variable for the moment. A mw_scanf()
wrapper might help and avoid repetitions. The overhead is negligible,
because the io code is non-critical for speed.