scripts: rename mw-foo.sh to foo.sh
The ambiguity between modules and scripts with similar names was handled
with capitalisation in megawave 3.01. This has to be changed for the
next versions, if we want to be able to run on case-insensitive
filesystems, like the common FAT filesystem on MS Windows.

The anbiguity will be managed now by explicitly calling the scripts with
their full names, including the ".sh" extension.