scripts: improve the installation procedure and wrapper
The wrapper will be used to call any megawave executable or script; this
is a convenient way to ensure the proper environment variables are
always set (LD_PRELOAD, path to data files, ....). It is a low overhead
method (1000 executions of the wrapper in 2 seconds).

The wrapper will emulate the previous mwdoc and cmw3 utilities; for
mwdoc, it uses html documentatioon files, because they are searchable,
and accessible even in a text-only environment.

Users can set their own configuration in ~/.config/megawave3.config.
System-wide settings are also possible in /etc/megawave3.config.